requirements.txt


numpy
Pillow
pandas==1.1.5
tensorflow==2.4.0
keras-tuner==1.0.2
opencv_contrib_python
tensorflow-datasets==4.0.1
tensorflow-addons==0.12.1
